LayoutTests/fast/url/path-url-expected.txt

Canonicalization of path URLs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canonicalize('javascript:') is 'javascript:'
FAIL canonicalize('JavaScript:Foo') should be javascript:Foo. Was JavaScript:Foo.
PASS successfullyParsed is true

TEST COMPLETE
